==> newFlowCohortExtractionPipeline/MGB/icd_neg/1_MGB_discharge_summary_filter.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import os
import re
import gc

### Discharge Summaries
 - Read in all note metadata CSV files and filter InpatientNoteTypeDSC column for only 'Discharge Summary'
 - Get a list of unique patient ID numbers within this (len = 69684)
 - Randomize the rows of the comprehensive dataframe (len = 539579)
 - Filter for the first instance of each patient ID (essentially a random discharge summary from that patient)
 - Keep track of: Note title, patient ID, date

# year = 2013
# dfs = []
# for i in range(0, 12):
#     df = pd.read_csv(f'/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/MGB_Deidentified_Notes_March12th2024/mgb_notes_{year + i}_metadata.csv')
#     df = df[df['InpatientNoteTypeDSC'] == 'Discharge Summary']
#     dfs.append(df)

output_dir = '/home/jsearle/bigDrive/NAX/NLP-SAH_identification/cohortExtractionPipeline/MGB/CSVs'

# Concatenate all filtered CSV files
all_files = [os.path.join(output_dir, f) for f in os.listdir(output_dir) if f.startswith('discharge_summaries_')]
discharge_summaries = pd.concat((pd.read_csv(f) for f in all_files), axis=0, ignore_index=True)

# Show number of total discharge summaries
print(len(discharge_summaries))

# Show example
discharge_summaries.head()

discharged_patients = set(discharge_summaries['BDSPPatientID'])
print(len(discharged_patients))

### Randomization
 - Setting seed (25) just to be able to reproduce results

rand_DS_comp = discharge_summaries.sample(frac=1, random_state=25).reset_index(drop=True)
rand_DS_comp.head()

 - Keep only the first instance of each patient ID (Should equal 69684)

rand_DS = rand_DS_comp.drop_duplicates(subset='BDSPPatientID', keep='first')
print(len(rand_DS))

### Combine with ICD table
 - I will be looking at all ICD codes within the month leading up to the date of the discharge summary
 - Check the format of the dates
 - Merge based on patient ID and date

# reader = ThunderReader('/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/thunderpack_icd_9_10_1m_MGB')
# key_length = len(list(reader.keys()))
# print(key_length)

# dfs = []
# for i in tqdm(range(1, key_length + 1)):
#     df = reader[f'ICD_partition_{i}']
#     df = df[df['BDSPPatientID'].isin(discharged_patients)]
#     dfs.append(df)

output_dir = '/home/jsearle/bigDrive/NAX/NLP-SAH_identification/cohortExtractionPipeline/MGB/CSVs'

all_files = [os.path.join(output_dir, f) for f in os.listdir(output_dir) if f.startswith('filtered_ICD_partition_')]
df = pd.concat((pd.read_csv(f) for f in all_files), axis=0, ignore_index=True)
print(len(df))

print(df['ShiftedContactDTS'][2])
print(type(df['ShiftedContactDTS'][2]))

 - Looks like the Contact Date column is full of strings
 - Converting to TimeStamp will make filtering easier

df['ShiftedContactDTS'] = pd.to_datetime(df['ShiftedContactDTS'])
print(df['ShiftedContactDTS'][2])
print(type(df['ShiftedContactDTS'][2]))

rand_DS.loc[:, 'ContactDate'] = pd.to_datetime(rand_DS['ContactDate'], format='%Y%m%d')
print(rand_DS['ContactDate'][2])
print(type(rand_DS['ContactDate'][2]))

merged_df = pd.merge(df, rand_DS, on='BDSPPatientID')

merged_df.head()

At this point I have merged the dataframes based on the patient ID
 - Both dates are listed (ICD date is 'ShiftedContactDTS' while note date is 'ContactDate')
 - Next step is to keep only the rows where the ContactDate is no more than one month after the ShiftedContactDTS

# merged_df_one_month = merged_df[(merged_df['ShiftedContactDTS'] >= merged_df['ContactDate'] - pd.DateOffset(months=1)) & 
#                                 (merged_df['ShiftedContactDTS'] <= merged_df['ContactDate'])]

# +/- one month
merged_df_one_month = merged_df[(merged_df['ShiftedContactDTS'] >= merged_df['ContactDate'] - pd.DateOffset(months=1)) & 
                                (merged_df['ShiftedContactDTS'] <= merged_df['ContactDate'] + pd.DateOffset(months=1))]


print(len(merged_df_one_month))
merged_df_one_month.head()

This df is getting a bit hectic
 - Let's remove the fluff and change some column names to be more readable
 - Columns to keep: BDSPPatientID, ShiftedContactDTS, ICDCD, code_type, ContactDate, DeidentifiedName
 - Rename these to: ID, ICD_Date, ICD, CodeType, NoteDate, NoteTitle
 - Move the dates next to each other

columns_to_keep = ['BDSPPatientID', 'ShiftedContactDTS', 'ICDCD', 'code_type', 'ContactDate', 'DeidentifiedName']
clean_df = merged_df_one_month[columns_to_keep]
clean_df.head()

rename_dict = {
    'ShiftedContactDTS': 'ICD_Date', 
    'ICDCD': 'ICD', 
    'code_type': 'CodeType', 
    'ContactDate': 'NoteDate', 
    'DeidentifiedName': 'NoteTitle'
}

clean_df = clean_df.rename(columns=rename_dict)
clean_df.head()

new_order = ['BDSPPatientID', 'ICD_Date', 'NoteDate', 'ICD', 'CodeType', 'NoteTitle']
clean_df = clean_df[new_order]
clean_df.head()

print(len(clean_df))

Now that the data has been processed a fair amount, lets store the df as a csv that we can just load up directly into another notebook (such as to filter based on cpt code association)

clean_df.to_csv('pre_cpt_filtering.csv', index=False, header=True, sep=',', na_rep='NA')

test = pd.read_csv('pre_cpt_filtering.csv')
test.head()
