==> newFlowCohortExtractionPipeline/MGB/icd_neg/2_MGB_cpt_filter.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re
import gc

# Goal of this notebook:
 - Filter my current dataframe (pre_cpt_filtering.csv) to only include patients who have either had a head CT or a brain MRI
 - Plan to do so:
   - Read in cpt df
   - Filter for head CT and brain MRI codes
   - Extract list of patient ID's
   - Read in pre_cpt_filtering.csv
   - Pull ID column, compare with cpt ID's, and keep overlap
   - Filter pre_cpt_filtering based on the new list


reader = ThunderReader('/home/jsearle/bigDrive/Dropbox/zz_EHR_Thunderpacks/MGB/thunderpack_cpt_1m_MGB')
key_length = len(list(reader.keys()))
print(key_length)

### CPT Codes to filter for:
**70450**: 	CT Head or Brain without contrast\
**70460**:	CT Head or Brain with contrast\
**70470**: 	CT Head or Brain without contrast, followed by contrast and further sections\
**70551**: 	MRI Brain (including brain stem) without contrast\
**70552**: 	MRI Brain (including brain stem) with contrast\
**70553**: 	MRI Brain (including brain stem) without contrast, followed by with contrast

code_regex = "(?:70450|70460|70470|70551|70552|70553)"

dfs = []
for i in tqdm(range(1, key_length + 1)):
    df = reader[f'cpt_partition_{i}']
    df = df[df['CPT'].astype(str).str.match(code_regex)]
    dfs.append(df)

df = pd.concat(dfs, axis=0, ignore_index=True)
print(len(df))

df.head()

unique_pt_ids = set(df['BDSPPatientID'])
print(len(unique_pt_ids))

The variable 'unique_pt_ids' is a set of all patient ID's associated with a patient that has had either a Head CT or Brain MRI\
Next Steps:
 - Read in the pre_cpt_filtering.csv dataframe
 - Pull the unique pt ID's from that df
 - Find the overlap

pre_cpt_filtering = pd.read_csv('pre_cpt_filtering.csv')
pre_cpt_filtering.head()

print(type(pre_cpt_filtering['BDSPPatientID'][0]))
print(type(df['BDSPPatientID'][0]))

pre_cpt_filtering['BDSPPatientID'] = pre_cpt_filtering['BDSPPatientID'].astype(int).astype(str)

pre_cpt_filtering.head()

unique_pre_cpt = set(pre_cpt_filtering['BDSPPatientID'])

overlap = unique_pt_ids.intersection(unique_pre_cpt)
print(len(overlap))

Looks like we have 23,899 patients who overlap.  Let's filter the pre_cpt file to only include those patients

prepped_df = pre_cpt_filtering[pre_cpt_filtering['BDSPPatientID'].isin(overlap)]
print(len(prepped_df))

prepped_df.head()

# Goal Accomplished
Summary:
 - Got the CPT dataframe read in while filtering for head CTs and brain MRIs
 - Pulled unique list of patient ID's
 - Got the pre_cpt_filtering csv file read in
 - Pulled unique list of patient ID's from this df
 - Found the overlap
 - Filtered the csv df to only include this overlap

Next:
 - Save it to a new csv
 - Create cohorts based on ICD code +/-

prepped_df.to_csv('prepped_df.csv', index=False, header=True, sep=',', na_rep='NA')

test = pd.read_csv('prepped_df.csv')
test.head()
