==> newFlowCohortExtractionPipeline/MGB/icd_neg/3_MGB_icd_filter.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import re
import gc

# Goal of this notebook:
 - Filter notes by ICD code indicating lack of SAH:
   - ICD -
 - Plan to do so:
   - For ICD +, filter using regex to include only ICD codes that I have listed
   - For ICD -, do the opposite
   - Remember that there are more rows than unique patients.  ie: some patients will have multiple ICD codes all linked to the same note
   - To fix that - try to condense rows into one per note by making the ICD column contain a list of ICD's (not tidy but I feel like it works here)

Start by reading in prepped_df.csv

prepped_df = pd.read_csv('prepped_df.csv')
prepped_df.head()

Let's repeat the process but matching everything except TBI codes to get our ICD- cohort

code_regex = "^(?!I60|430).*"
icd_neg = prepped_df[prepped_df['ICD'].astype(str).str.match(code_regex, flags=re.I)]
icd_neg.head()

icd_neg.loc[:, 'ICD'] = icd_neg['ICD'].astype(str)
grouped = icd_neg.groupby('NoteTitle')['ICD'].apply(lambda x: ', '.join(x)).reset_index()
icd_neg_unique = icd_neg.drop_duplicates(subset=['NoteTitle']).drop(columns=['ICD'])
icd_neg_unique = pd.merge(icd_neg_unique, grouped, on='NoteTitle', how='left')

print(len(icd_neg_unique))

Finally, let's choose our sample of 1087.  Same size as ICD +

cohort_neg = icd_neg_unique.sample(n=1087, replace=False, random_state=1)
print(len(cohort_neg))

cohort_neg.head()

I guess the ICD code really only matters for the plus cohort in case we wanted to look into the spread of certain TBI codes

cohort_neg = cohort_neg.drop(columns=['ICD_Date', 'CodeType', 'ICD'])
cohort_neg.head()

cohort_neg.to_csv('cohort_neg.csv', index=False, header=True, sep=',', na_rep='NA')

test = pd.read_csv('cohort_neg.csv')
test.head()

# Goal Accomplished
Summary:
 - Used regex to search for ICD code specifics
 - Got the unique list of note titles
 - Created cohorts of 1000 (randomly)
 - Kept important columns and exported to CSV
 
Next:
 - Bram said I need to get the whole note text into a new column somehow
