==> newFlowCohortExtractionPipeline/MGB/icd_neg/4_MGB_note_reading.ipynb <==
import pandas as pd
from thunderpack import ThunderReader
from tqdm import tqdm
import tarfile
import zipfile
import random
import os
import shutil
import re
tqdm.pandas()

# Goal of this notebook:
 - Add a column to the dataframe which includes the text of the note associated with the row
 - Plan:
    - Figure out how to open tar archive from here
    - Create a function that opens the tar archive, finds the file, and extracts the text
    - Make a pandas series from these text blocks and append to the initial dataframe

cohort_neg = pd.read_csv('cohort_neg.csv')
cohort_neg.head()

notes = cohort_neg['NoteTitle']
print(notes[0])

Before continuing, we will make some helper functions that will make the process easier.
 - In order to go through the whole list of notes and extract their text, we will need to know which zip file they are
 - To do this we will use regex to extract the year and append it to the base file path
 - We also need to modify the read_from_zip method

# This will extract the year from the file name
def extract_year(file_name):
    match = re.search(r'_(\d{8})\.txt$', file_name)
    if match:
        return match.group(1)[:4]
    return None

def read_file_from_zip(base_path, file_name):
    year = extract_year(file_name)
    if year:
        zip_path = f'{base_path}/mgb_notes_{year}.zip'
        try:
            with zipfile.ZipFile(zip_path, 'r') as zipf:
                with zipf.open(f'{year}/{file_name}') as file:
                    content = file.read()
                    return content.decode('utf-8', errors = 'ignore')
        except (KeyError, FileNotFoundError):
            return None
    return None

With these updated helper functions we should be able to attach the txt of each note to the corresponding rows

base_path = '/home/jsearle/bigDrive/NAX/zipNotes/MGB'

cohort_neg['text'] = cohort_neg['NoteTitle'].progress_apply(lambda x: read_file_from_zip(base_path, x))

cohort_neg.to_csv('cohort_plus_text.csv', index=False, header=True, sep=',', na_rep='NA')

cohort_neg.head()

I forgot to check the character counts -- let's make sure each note has more than 500 characters

def check_word_count(text, threshold):
    return len(text) > threshold

threshold = 500
cohort_neg['valid'] = cohort_neg['text'].apply(lambda x: check_word_count(x, threshold))

cohort_neg.head()

print(sum(cohort_neg['valid']))

false_indeces = cohort_neg.index[~cohort_neg['valid']].tolist()
print(false_indeces)

cohort_neg = cohort_neg.drop(index=false_indeces)

print(len(cohort_neg))


# cohort_neg = cohort_neg.drop(columns=['ICD_Date', 'CodeType'])
print(len(cohort_neg))
cohort_neg.head()

Double check that the character count is valid

threshold = 500
cohort_neg['valid'] = cohort_neg['text'].apply(lambda x: check_word_count(x, threshold))
print(sum(cohort_neg['valid']))

updated_cohort_neg = cohort_neg.drop(columns=['valid'])
updated_cohort_neg.to_csv('final_icd_neg_df.csv', index=False, header=True, sep=',', na_rep='NA')

test = pd.read_csv('final_icd_neg_df.csv')
test.head()
